+++ variables.tf
@@ -63,13 +63,13 @@ variable "custom_vwan_name" {
   default     = null
 }
 
-variable "disable_vpn_encryption" {
-  description = "Boolean flag to specify whether VPN encryption is disabled"
+variable "vpn_encryption_enabled" {
+  description = "Boolean flag to specify whether VPN encryption is enabled"
   type        = bool
-  default     = false
+  default     = true
 }
 
-variable "allow_branch_to_branch_traffic" {
+variable "branch_to_branch_traffic_allowed" {
   description = "Boolean flag to specify whether branch to branch traffic is allowed"
   type        = bool
   default     = true
@@ -81,43 +81,47 @@ variable "office365_local_breakout_category" {
   default     = "None"
 }
 
-variable "vwan_type" {
+variable "virtual_wan_type" {
   description = "Specifies the Virtual WAN type. Possible Values include: `Basic` and `Standard`"
   type        = string
   default     = "Standard"
 }
 
-variable "vwan_extra_tags" {
+variable "virtual_wan_extra_tags" {
   description = "Extra tags for this virtual wan"
   type        = map(string)
   default     = {}
 }
 
 # Virtual Hub specific variables
-variable "custom_vhub_name" {
+variable "custom_virtual_hub_name" {
   description = "Custom virtual hub's name"
   type        = string
   default     = null
 }
 
-variable "vhub_address_prefix" {
+variable "virtual_hub_address_prefix" {
   description = "The address prefix which should be used for this virtual hub. Cannot be smaller than a /24. A /23 is recommended by Azure"
   type        = string
+  validation {
+    condition     = tonumber(split("/", var.virtual_hub_address_prefix)[1]) <= 24
+    error_message = "Virtual Hub address prefix must be at least /24. A /23 is recommended by Azure."
+  }
 }
 
-variable "vhub_sku" {
+variable "virtual_hub_sku" {
   description = "The sku of the virtual hub. Possible values are `Basic` and `Standard`"
   type        = string
   default     = "Standard"
 }
 
-variable "vhub_extra_tags" {
+variable "virtual_hub_extra_tags" {
   description = "Extra tags for this virtual hub"
   type        = map(string)
   default     = {}
 }
 
-variable "vhub_routes" {
+variable "virtual_hub_routes" {
   description = "List of route blocks. next_hop_ip_address values can be azure_firewall or an ip address"
   type = list(object({
     address_prefixes    = list(string),
@@ -127,53 +131,53 @@ variable "vhub_routes" {
 }
 
 # Express route variables
-variable "enable_express_route" {
+variable "express_route_enabled" {
   description = "Enable or not express route configuration"
   type        = bool
   default     = false
 }
-variable "custom_ergw_name" {
+variable "custom_express_route_gateway_name" {
   description = "Custom express route gateway name"
   type        = string
   default     = null
 }
 
-variable "ergw_exta_tags" {
+variable "express_route_gateway_exta_tags" {
   description = "Extra tags for Express Route Gateway"
   type        = map(string)
   default     = {}
 }
 
-variable "custom_erc_name" {
+variable "custom_express_route_circuit_name" {
   description = "Custom express route circuit name"
   type        = string
   default     = null
 }
 
-variable "er_scale_unit" {
+variable "express_route_gateway_scale_unit" {
   description = "The number of scale unit with which to provision the ExpressRoute gateway."
   type        = number
   default     = 1
 }
-variable "erc_peering_location" {
-  description = "The name of the peering location that this Express Route is."
+variable "express_route_circuit_peering_location" {
+  description = "Express route peering location."
   type        = string
   default     = null
 }
 
-variable "erc_bandwidth_in_mbps" {
+variable "express_route_circuit_bandwidth_in_mbps" {
   description = "The bandwith in Mbps of the circuit being created on the Service Provider"
   type        = number
   default     = null
 }
 
-variable "erc_service_provider" {
+variable "express_route_circuit_service_provider" {
   description = "The name of the ExpressRoute Service Provider."
   type        = string
   default     = null
 }
 
-variable "er_sku" {
+variable "express_route_sku" {
   description = "ExpressRoute SKU"
   type = object({
     tier   = string,
@@ -185,92 +189,92 @@ variable "er_sku" {
   }
 }
 
-variable "enable_er_private_peering" {
+variable "express_route_private_peering_enabled" {
   description = "Enable Express Route Circuit Private Peering"
   type        = bool
   default     = false
 }
 
-variable "erc_private_peering_primary_peer_address_prefix" {
+variable "express_route_circuit_private_peering_primary_peer_address_prefix" {
   description = "Primary peer address prefix for Express Route Circuit private peering"
   type        = string
   default     = null
 }
 
-variable "erc_private_peering_secondary_peer_address_prefix" {
+variable "express_route_circuit_private_peering_secondary_peer_address_prefix" {
   description = "Secondary peer address prefix for Express Route Circuit private peering"
   type        = string
   default     = null
 }
 
-variable "erc_private_peering_shared_key" {
+variable "express_route_circuit_private_peering_shared_key" {
   description = "Shared secret key for Express Route Circuit Private Peering"
   type        = string
   default     = null
 }
 
-variable "erc_private_peering_vlan_id" {
+variable "express_route_circuit_private_peering_vlan_id" {
   description = "VLAN Id for Express Route "
   type        = number
   default     = null
 }
 
-variable "erc_private_peering_peer_asn" {
+variable "express_route_circuit_private_peering_peer_asn" {
   description = "Peer BGP ASN for Express Route Circuit Private Peering"
   type        = number
   default     = null
 }
 
 # Firewall specific variables
-variable "enable_firewall" {
+variable "firewall_enabled" {
   description = "Enable or not Azure Firewall in the Virtual Hub"
   type        = bool
   default     = true
 }
 
-variable "custom_fw_name" {
+variable "custom_firewall_name" {
   description = "Custom firewall's name"
   type        = string
   default     = null
 }
 
-variable "fw_extra_tags" {
+variable "firewall_extra_tags" {
   description = "Extra tags for Firewall resource"
   type        = map(string)
   default     = {}
 }
 
-variable "fw_sku_tier" {
+variable "firewall_sku_tier" {
   description = "Sku tier of the Firewall. Possible values are `Premium` and `Standard`."
   type        = string
   default     = "Standard"
 }
 
-variable "fw_policy_id" {
+variable "firewall_policy_id" {
   description = "ID of the Firewall Policy applied to this Firewall."
   type        = string
   default     = null
 }
 
-variable "fw_availibility_zones" {
+variable "firewall_availibility_zones" {
   description = " availability zones in which the Azure Firewall should be created."
   type        = list(number)
   default     = [1, 2, 3]
 }
 
-variable "fw_public_ip_count" {
+variable "firewall_public_ip_count" {
   description = "Number of public IPs to assign to the Firewall."
   type        = number
   default     = 1
 }
 
-variable "fw_dns_servers" {
+variable "firewall_dns_servers" {
   description = "List of DNS servers that the Azure Firewall will direct DNS traffic to for the name resolution"
   type        = list(string)
   default     = null
 }
 
-variable "fw_private_ip_ranges" {
+variable "firewall_private_ip_ranges" {
   description = "List of SNAT private CIDR IP ranges, or the special string `IANAPrivateRanges`, which indicates Azure Firewall does not SNAT when the destination IP address is a private range per IANA RFC 1918"
   type        = list(string)
   default     = null